reorg(r9e): enforce the layering contracts with import-linter (#489)
The two-tier rule has been written down since R5 and had 20 live
counterexamples by R9. Prose does not hold a layer; this makes it
mechanical. .importlinter carries three contracts and CI runs
lint-imports without continue-on-error, unlike the advisory lint steps
beside it.

Clearing the last violations so it passes with no exemptions:

- RedisDedupSet was in services/daemon/ but read by core/ingestion and
  core/federation. It is ingestion machinery -> core/ingestion/dedup.py.
  Federation reaching it is the direction CONTEXT.md already sanctions.
- KafkaConfig was in services/daemon/config.py but read by
  core/ingestion. -> core/ingestion/kafka_config.py, re-exported for
  DaemonConfig.
- ollama_ping was in core/llm/providers/discovery.py with the supervisor
  as its only caller. A process-liveness probe is mechanism, so it moves
  into the supervisor and leaves core/llm.
- _sync_bifrost is the opposite case: genuinely LLM knowledge sitting in
  a platform module. The supervisor now exposes a post_start_sync hook
  defaulting to a no-op, core/llm/bifrost/admin.py provides
  sync_after_ollama_start, and the composition roots wire the two — the
  local-services router for the API path, and a new
  scripts/ollama_supervise.py for the CLI path, which replaces the
  supervisor's own main() so no core.llm reference remains. scripts/lib.sh
  invokes the script instead of `python3 -m core.llm.providers.ollama`,
  which r9c had already left dangling.

Contracts: 3 kept, 0 broken, over 264 files and 735 dependencies.

CI also still linted backend/ (gone since r9a) and measured coverage
with --cov=backend and --cov=daemon, neither of which is a package.
Now services/ and core/.

tests/security/ ran only in nightly, so a genuine auth bypass has been
sitting red unnoticed: POST /api/integrations/vstrike/findings answers
200 with no bearer, though main.py lists it public on the grounds that
it "uses its own bearer API-key dependency". It is now in the gating
job, with that one case excluded by -k so it stays visible rather than
silently unrun. Fixing it is separate work.

tests/llm/test_boundary.py is the hand-rolled precursor to this gate;
its UPWARD_PACKAGES now names services, and its grandfathered exemption
for backend/schemas/tool_schemas.py is dropped — r9a moved that file
into core/llm/, so the import is no longer upward.

Unit+llm+security 1294 passed / 4 failed — the same four pre-existing
failures.
