+++ .github/workflows/ci-cd.yml
@@ -39,23 +39,28 @@ jobs:
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
-          pip install flake8 black isort mypy
+          pip install flake8 black isort mypy import-linter
 
       - name: Run flake8
-        run: flake8 backend/ services/ core/ --max-line-length=120 --exclude=venv,__pycache__
+        run: flake8 services/ core/ --max-line-length=120 --exclude=venv,__pycache__
         continue-on-error: true
       
       - name: Run black
-        run: black --check backend/ services/ core/ --exclude=venv
+        run: black --check services/ core/ --exclude=venv
         continue-on-error: true
       
       - name: Run isort
-        run: isort --check-only backend/ services/ core/
+        run: isort --check-only services/ core/
         continue-on-error: true
       
       - name: Run mypy
-        run: mypy backend/ services/ --ignore-missing-imports
+        run: mypy services/ core/ --ignore-missing-imports
         continue-on-error: true
+
+      # The layering contracts in .importlinter (epic #481). Deliberately no
+      # continue-on-error: the rest of this job is advisory, this one gates.
+      - name: Run import-linter
+        run: lint-imports
 
   lint-frontend:
     name: Lint Frontend
@@ -124,7 +129,13 @@ jobs:
           # This step also carries the #415 ratchets (test_no_ambient_state,
           # test_settings_env_example): no continue-on-error here, unlike the lint
           # job, so a new raw env read or env.example drift fails the pipeline.
-          pytest tests/unit/ tests/llm/ -m "not external_service" -v --cov=backend --cov=services --cov=daemon --cov=core --cov-report=xml --cov-report=term
+          # tests/security/ previously ran only in nightly, so a genuine
+          # auth-bypass sat red and unnoticed: POST /api/integrations/vstrike/findings
+          # answers 200 without a bearer. Gate the other 83 now; the bypass is
+          # deselected so it stays visible here rather than silently unrun.
+          pytest tests/unit/ tests/llm/ tests/security/ -m "not external_service" -v \
+            -k "not test_vstrike_inbound_without_bearer_uses_api_key_gate" \
+            --cov=services --cov=core --cov-report=xml --cov-report=term
         env:
           TESTING: "true"
           CLAUDE_API_KEY: ${{ secrets.CLAUDEAPI }}
@@ -218,7 +229,7 @@ jobs:
 
       - name: Run integration tests
         run: |
-          pytest tests/integration/ -v --cov=backend --cov-report=xml
+          pytest tests/integration/ -v --cov=services --cov=core --cov-report=xml
         env:
           CLAUDE_API_KEY: ${{ secrets.CLAUDEAPI }}
           ANTHROPIC_API_KEY: ${{ secrets.CLAUDEAPI }}
