+++ core/federation/runner.py
@@ -26,7 +26,7 @@ from datetime import datetime
 from typing import Any, Dict, Optional
 
 from core.config import DEFAULT_REDIS_URL, get_settings
-from services.daemon.dedup import RedisDedupSet
+from core.ingestion.dedup import RedisDedupSet
 from core.federation import registry, store
 from core.federation.seed import seed_federation_sources
