+++ core/ingestion/kafka_consumer_service.py
@@ -24,8 +24,8 @@ import logging
 from datetime import datetime
 from typing import Any, Dict, Optional
 
-from services.daemon.config import KafkaConfig
-from services.daemon.dedup import RedisDedupSet
+from core.ingestion.kafka_config import KafkaConfig
+from core.ingestion.dedup import RedisDedupSet
 
 logger = logging.getLogger(__name__)
