+++ tests/unit/test_kafka_consumer.py
@@ -12,8 +12,8 @@ import asyncio
 import json
 import pytest
 
-from services.daemon.config import KafkaConfig
-from services.daemon.dedup import RedisDedupSet
+from core.ingestion.kafka_config import KafkaConfig
+from core.ingestion.dedup import RedisDedupSet
 from core.ingestion.kafka_consumer_service import KafkaConsumerService
